requirements.txt
pandas==1.3.0
python-docx==0.8.11
xlrd==2.0.1
XlsxWriter==1.4.4
